# config/openapi.yaml
openapi: 3.0.0
info:
  title: TFT Bible API
  description: API for Teamfight Tactics composition tracking and trait analysis
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.tftbible.com/api/v1
    description: Production server
paths:
...
  /champions:
    get:
      summary: Get all champions
      description: Retrieve a list of all available champions
      parameters:
        - name: limit
          in: query
          description: Maximum number of results to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: set
          in: query
          description: Filter by TFT set
          required: false
          schema:
            type: string
        - name: cost
          in: query
          description: Filter by champion cost (1-5 stars)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: traits
          in: query
          description: Filter by comma-separated list of traits
          required: false
          schema:
            type: string
        - name: search
          in: query
          description: Search by champion name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of champions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedChampions'
        '400':
          $ref: '#/components/responses/BadRequest'
  /champions/{id}:
    get:
      summary: Get champion by ID
      description: Retrieve details of a specific champion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Champion ID
      responses:
        '200':
          description: Champion details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Champion'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
    Champion:
      type: object
      properties:
        id:
          type: string
          example: "TFT15_Ahri"
        name:
          type: string
          example: "Ahri"
        cost:
          type: integer
          minimum: 1
          maximum: 5
          example: 2
        traits:
          type: array
          items:
            type: string
          example: ["Shapeshifter", "Sorcerer"]
        stats:
          $ref: '#/components/schemas/ChampionStats'
        ability:
          $ref: '#/components/schemas/ChampionAbility'
        image:
          type: string
          format: uri
          example: "https://ddragon.leagueoflegends.com/cdn/14.23.1/img/champion/Ahri.png"
      required:
        - id
        - name
        - cost
        - traits
    ChampionStats:
      type: object
      properties:
        health:
          type: number
          format: float
          example: 650.0
        mana:
          type: number
          format: float
          example: 80.0
        starting_mana:
          type: number
          format: float
          example: 30.0
        armor:
          type: number
          format: float
          example: 20.0
        magic_resist:
          type: number
          format: float
          example: 20.0
        attack_damage:
          type: number
          format: float
          example: 40.0
        attack_speed:
          type: number
          format: float
          example: 0.65
        attack_range:
          type: number
          format: float
          example: 3.0
        crit_chance:
          type: number
          format: float
          example: 0.25
        crit_multiplier:
          type: number
          format: float
          example: 1.5
    ChampionAbility:
      type: object
      properties:
        name:
          type: string
          example: "Orb of Deception"
        description:
          type: string
          example: "Ahri fires an orb in a line that deals damage on the way out and heals her on the way back."
        type:
          type: string
          example: "Active"
        targeting:
          type: string
          example: "Enemies"
        damage_type:
          type: string
          example: "Magic"
...
    PaginatedChampions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Champion'
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 8
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters or missing required fields
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Invalid request parameters"
              errors:
                type: array
                items:
                  type: string
                example: ["limit must be between 1 and 100", "invalid trait specified"]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: "Resource not found"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT